move parse_xtal to fragalysis.py
