+++ asapdiscovery/docking/scripts/find_all_mcs.py
@@ -8,7 +8,7 @@ import sys
 
 sys.path.append(f"{os.path.dirname(os.path.abspath(__file__))}/../")
 from asapdiscovery.data.utils import load_exp_from_sdf
-from asapdiscovery.docking.docking import parse_xtal
+from asapdiscovery.data.fragalysis import parse_xtal
 from asapdiscovery.docking.mcs import (
     rank_structures_openeye,
     rank_structures_rdkit,
