+++ asapdiscovery/docking/scripts/find_mcs_mers.py
@@ -5,7 +5,7 @@ import os
 import sys
 
 sys.path.append(f"{os.path.dirname(os.path.abspath(__file__))}/../")
-from asapdiscovery.docking.docking import parse_xtal
+from asapdiscovery.data.fragalysis import parse_xtal
 from asapdiscovery.docking.mcs import (
     rank_structures_openeye,
     rank_structures_rdkit,
